bilinear_zero_shot_model/requirements.txt
lightning==2.0.9
matplotlib==3.8.0
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0.1
scikit_learn==1.4.0
torch==1.13.1
wandb==0.15.10